+++ lib/screens/debug/sentry_flutter_screen.dart
@@ -41,10 +41,7 @@ class _SentryFlutterScreenState extends State<SentryFlutterScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: Text('Sentry Flutter SDK - Debug'),
-          brightness: Brightness.dark,
-        ),
+        appBar: AppBar(title: Text('Sentry Flutter SDK - Debug')),
         body: SingleChildScrollView(
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,