+++ lib/screens/license/license_screen.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:sentry_mobile/utils/sentry_colors.dart';
 
 class LicenseScreen extends StatelessWidget {
@@ -11,8 +10,6 @@ class LicenseScreen extends StatelessWidget {
     final ThemeData themeData = Theme.of(context);
     return Theme(
         data: themeData.copyWith(
-          appBarTheme: AppBarTheme(
-              brightness: Brightness.dark, backgroundColor: SentryColors.rum),
           textTheme: themeData.textTheme.copyWith(
             subtitle1: TextStyle(
               fontSize: 16,